+++ example/main.js
@@ -2,12 +2,12 @@
  * Created by nocoolyoyo on 2018/10/18.
  */
 import { initClient, isStyleUnit, RegexMap,
-	     scrollToXView, scrollToYView, scrollToX, scrollToY, scrollTo, getClient } from '../src/index'
+	     scrollToXView, scrollToYView, scrollToX, scrollToY, scrollTo, getClient, copy2Board, appendUrlParams } from '../src/index'
 console.log(RegexMap)
 console.log(isStyleUnit('400px'))
 
 
-initClient()
+
 
 document.querySelector('.scroller').scrollTop = 9000
 document.querySelector('.scroller').scrollLeft = 9000
@@ -25,18 +25,31 @@ setTimeout(()=> {
 	// 	justify: 0
 	// })
 	scrollToX(document.querySelector('.scroller'), {
-		target: document.querySelector('.target'),
-		position: 'end',
+		target: 200,
+		position: 'top',
 		justify: 0
 	})
 	scrollToY(document.querySelector('.scroller'), {
-		target: document.querySelector('.target'),
-		position: 'end',
+		target: 200,
+		position: 'top',
 		justify: 0
 	})
 
 }, 500)
+
+setInterval(()=>{
+	initClient()
+},1000)
 console.log(getClient())
+
+console.log(appendUrlParams({
+	haha: 1
+}))
+document.querySelector('#clipboard').addEventListener('click', function () {
+	copy2Board('哈哈哈哈哈哈哈')
+})
+
+