+++ src/compare.js
@@ -1,0 +1,8 @@
+export function isPreVersion(preVersion = '', sufVersion = '') {
+	if (preVersion === '') {
+		return true
+	}
+
+	const version
+	return Client
+}