+++ src/urlParams.js
@@ -1,4 +1,28 @@
 /**
+ * @function 获取URL全部参数值
+ * @param name
+ * @return {string}
+ *
+ * @author: nocoolyoyo
+ * @date: 2018-03-11
+ */
+export function getUrlParams(url = window.location.href) {
+	const _url = decodeURI(url.replace(/\+/g, '%20'))
+  console.log(_url)
+  const result = {}
+  let queryArray = url.split('?')
+  if (queryArray.length <= 1) {
+    return result
+  }
+	queryArray = queryArray[1].split('&')
+	queryArray.forEach(query =>  {
+	  const queryKVArray= query.split('=')
+    console.log(queryKVArray)
+		result[queryKVArray[0]] = queryKVArray[1]
+  })
+	return result
+}
+/**
  * @function 获取指定的URL参数值
  * @param name
  * @return {string}
@@ -17,7 +41,7 @@ export function getUrlParam(name, url = window.location.href) {
 }
 
 /**
- * @function 追加指定的URL参数值
+ * @function 追加指定的URL参数值,存在会覆盖
  * @param params
  * @param url
  * @return {string}
@@ -26,14 +50,14 @@ export function getUrlParam(name, url = window.location.href) {
  * @date: 2018-09-04
  */
 export function appendUrlParams(params = {}, url = window.location.href) {
+  const oriParams = getUrlParams(url)
+	const newParams = Object.assign(oriParams, params)
   // url已存在参数
   let tempStr = ''
-  Object.keys(params).forEach(key => {
+  Object.keys(newParams).forEach(key => {
     tempStr += `&${key}=${params[key]}`
   })
-  if (url.indexOf('?')) {
-    tempStr = tempStr.replace(/\?/, '&')
-  }
+	tempStr = tempStr.replace(/&/, '?')
   url += tempStr
   return url
 }